+++ modules/core/src/main/java/io/github/palexdev/mfxcore/observables/OnChanged.java
@@ -219,6 +219,7 @@ public class OnChanged<T> extends When<T> {
 			observableValue.removeListener(listener);
 			listener = null;
 			whens.remove(observableValue);
+			observableValue = null;
 		}
 	}
 }