+++ modules/core/src/main/java/io/github/palexdev/mfxcore/observables/OnInvalidated.java
@@ -221,6 +221,7 @@ public class OnInvalidated<T> extends When<T> {
 			observableValue.removeListener(listener);
 			listener = null;
 			whens.remove(observableValue);
+			observableValue = null;
 		}
 	}
 }