+++ modules/core/src/test/java/unit/observables/WhenTests.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (C) 2023 Parisi Alessandro - alessandro.parisi406@gmail.com
+ * This file is part of MaterialFX (https://github.com/palexdev/MaterialFX)
+ *
+ * MaterialFX is free software: you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public License
+ * as published by the Free Software Foundation; either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * MaterialFX is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+ * See the GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with MaterialFX. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package unit.observables;
+
+import io.github.palexdev.mfxcore.observables.OnInvalidated;
+import io.github.palexdev.mfxcore.observables.When;
+import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleIntegerProperty;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+public class WhenTests {
+
+	@Test
+	void testMultipleFail() {
+		IntegerProperty ip = new SimpleIntegerProperty();
+
+		When.onInvalidated(ip)
+				.then(System.out::println)
+				.listen();
+
+		assertThrows(
+				IllegalArgumentException.class,
+				() -> When.onInvalidated(ip)
+						.then(n -> System.out.println("This should not happen"))
+						.listen()
+		);
+	}
+
+	@Test
+	void testMultipleForce() {
+		IntegerProperty ip = new SimpleIntegerProperty();
+
+		OnInvalidated<Number> first = When.onInvalidated(ip)
+				.then(System.out::println)
+				.listen();
+
+		assertDoesNotThrow(
+				() -> {
+					When.onInvalidated(ip)
+							.then(n -> System.out.println("This should not happen"))
+							.forceListen();
+				}
+		);
+		assertTrue(first.isDisposed());
+	}
+}